// code/src/Enums/NewsCategories.ts
export enum ArticleCategories {
    Binnenland = 'Binnenland',
    Buitenland = 'Buitenland',
    Cultuur = 'Cultuur & Media',
    Economie = 'Economie',
    Koningshuis = 'Koningshuis',
    Opmerkelijk = 'Opmerkelijk',
    Politiek = 'Politiek',
    Tech = 'Tech',
    Regionaal = 'Regionaal nieuws',
}